Unified MemoryUtils into Utils.
